Try to deploy as vercel function
