+++ api/resolvers/SeriesResolver.js
